+++ pymcuprog/deviceinfo/devices/attiny427.py
@@ -68,7 +68,8 @@ DEVICE_INFO = {
     # user_row
     'user_row_address_byte': 0x1300,
     'user_row_size_bytes': 0x20,
-    'user_row_page_size_bytes': 0x20,
+    # user_row is implemented as EEPROM so writes should not be padded to start of page, i.e. set page size to 1
+    'user_row_page_size_bytes': 0x01,
     'user_row_read_size_bytes': 0x01,
     'user_row_write_size_bytes': 0x01,
     'user_row_chiperase_effect': ChiperaseEffect.NOT_ERASED,